requirements.txt
aiohttp==1.0.5
async-timeout==3.0.0
certifi==2018.8.24
chardet==3.0.4
discord.py==0.16.12
idna==2.7
multidict==4.4.1
python-dotenv==0.9.1
requests==2.19.1
urllib3==1.23
websockets==3.4
